# backend/api-document/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Metrade API",
    "description": "This API is used for managing users, products, orders, order items, carts, cart items, token.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "schemas": {
...
      "Product": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Laptop"
          },
          "description": {
            "type": "string",
            "example": "A powerful laptop for developers."
          },
          "price": {
            "type": "number",
            "example": 1200
          },
          "stock_quantity": {
            "type": "number",
            "example": 10
          },
          "pickup_point": {
            "type": "string",
            "enum": [
              "Myllypuro",
              "Karamalmi",
              "Myyrmäki"
            ],
            "example": "Myllypuro"
          },
          "status": {
            "type": "string",
            "enum": [
              "processing",
              "active",
              "sold"
            ],
            "example": "active"
          }
        }
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Electronics",
            "description": "The name of the category"
          },
          "description": {
            "type": "string",
            "example": "Category for all electronic products",
            "description": "Description of the category"
          },
          "parent_id": {
            "type": "string",
            "example": "60d5f489f9e7c4a5d37fbe28",
            "description": "Reference to the parent category ID"
          },
          "ancestors": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "References to all ancestor categories"
            },
            "example": [
              "60d5f489f9e7c4a5d37fbe27",
              "60d5f489f9e7c4a5d37fbe26"
            ],
            "description": "Array of ancestor categories"
          },
          "children": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "References to direct child categories"
            },
            "example": [
              "60d5f489f9e7c4a5d37fbe29",
              "60d5f489f9e7c4a5d37fbe30"
            ],
            "description": "Array of direct child categories"
          },
          "is_active": {
            "type": "boolean",
            "example": true,
            "description": "Indicates whether the category is active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2024-10-04T08:40:51.620Z",
            "description": "Category creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2024-10-04T09:00:51.620Z",
            "description": "Category update timestamp"
          }
        }
      },
...
    },
    "securitySchemes": {
      "accessTokenCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      },
      "refreshTokenCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken"
      }
    }
  },
  "security": [
    {
      "accessTokenCookie": [],
      "refreshTokenCookie": []
    }
  ],
  "paths": {
...
    "/token/get-access-token": {
      "post": {
        "tags": [
          "Token"
        ],
        "summary": "Get a new access token using a refresh token",
        "security": [
          {
            "accessTokenCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "New access token generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "New access token generated"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "670154eda0f091fa500887ea"
                        },
                        "first_name": {
                          "type": "string",
                          "example": "John"
                        },
                        "last_name": {
                          "type": "string",
                          "example": "Doe"
                        },
                        "email": {
                          "type": "string",
                          "example": "john.doe@example.com"
                        },
                        "role": {
                          "type": "string",
                          "example": "user"
                        },
                        "photo_url": {
                          "type": "string",
                          "example": "/uploads/avatar.jpg"
                        },
                        "is_verified": {
                          "type": "boolean",
                          "example": true
                        },
                        "phone": {
                          "type": "string",
                          "example": "+1234567890"
                        },
                        "balance": {
                          "type": "number",
                          "example": 100.5
                        },
                        "card_id": {
                          "type": "string",
                          "example": "670154eda0f091fa500887ea"
                        },
                        "token_expired_at": {
                          "type": "integer",
                          "example": 1693779134000
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No refresh token provided or invalid/expired refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "No refresh token provided"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/categories/main-category/main-relationship": {
      "get": {
        "tags": [
          "Category"
        ],
        "security": [],
        "summary": "Get main categories with hierarchy",
        "operationId": "categoryHierarchy",
        "responses": {
          "200": {
            "description": "List of categories with hierarchy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No main categories found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/categories/get-children-categories/{categoryId}": {
      "get": {
        "tags": [
          "Category"
        ],
        "security": [],
        "summary": "Get child categories and products",
        "operationId": "getChildrenProducts",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the category to retrieve children for"
          }
        ],
        "responses": {
          "200": {
            "description": "List of products in the category and its children",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}